+++ uniffi_bindgen/src/bindings/kotlin/templates/Types.kt
@@ -163,6 +163,9 @@ object NoHandle
 {%- when Type::Map { key_type, value_type } %}
 {% include "MapTemplate.kt" %}
 
+{%- when Type::Set { inner_type } %}
+{% include "SetTemplate.kt" %}
+
 {%- when Type::CallbackInterface { module_path, name } %}
 {% include "CallbackInterfaceTemplate.kt" %}